changed schema based on barrys updates

+++ src/main/java/uk/gov/hmcts/darts/courthouse/service/CourthouseServiceImpl.java
@@ -28,7 +28,7 @@ public class CourthouseServiceImpl implements CourthouseService {
         Courthouse originalEntity = repository.getReferenceById(id);
 
         originalEntity.setCourthouseName(courthouse.getCourthouseName());
-        originalEntity.setCode(courthouse.getCode());
+        originalEntity.setCode(courthouse.getCode().shortValue());
 
         return repository.saveAndFlush(originalEntity);
     }
